Finish the rebase
